pyproject.toml
[project]
name = "engineering-experiment-with-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.65.0",
    "groq>=1.0.0",
    "ipykernel>=7.2.0",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]

[tool.uv.workspace]
members = [
    "apps/api",
]
